mobilegaze/requirements.txt
numpy
opencv-python
torch
pyyaml
uniface
cjm-byte-track
